Fix race condition by making check and mutation atomic as suggested by Lion. Also added error handling to reconstruction failure.
